Namespace BenString into its own module BenjaminString in order to prevent namespace pollution. BenString is still provided for compatibility reasons.
Some will probably suggest changing the name of the project from BenString to
BenjaminString. I propose we should not change the name of the project at this
time in order to reduce confusion amongst current users.

Regardless (or even irregardless), I suggest an FAQ be created to address the
many questions that will be raised by this very much needed and important
change.